Divided post and community into multiple files
